requirements.txt
# Core
django==3.1.3
django-configurations==2.2
pq==1.9.0
gunicorn==20.0.4

# Database
psycopg2==2.8.6
dj-database-url==0.5

# Email
dj-email-url==1.0.1

# Auth
social-auth-app-django==4.0.0

# Model Tools
django-simple-activity==1.1.0
django-money==1.1

# Rest apis
djangorestframework==3.12.2
djangorestframework-jsonapi==4.0.0

django-cors-headers==3.5.0
django-filter==2.4.0
querystring-parser==1.2.4

# Misc
iso3166==1.0.1
filetype==1.0.7
sentry-sdk==0.19.4
